Reorganized some things.
