+++ alchemy_framework/src/graphics/app_window.rs
@@ -1,44 +1,32 @@
 use winit::dpi::LogicalSize;
-use winit::event::{Event, VirtualKeyCode};
 use winit::event_loop::{ControlFlow, EventLoop};
 use winit::window::WindowBuilder;
-use winit_input_helper::WinitInputHelper;
-use winit::{ event::*};
 use futures;
 
-use crate::graphics;
+use super::*;
 
-///this is a function that takes no parameters and returns nothing, that can be used to get around
-///rust limits on inherientance that I would normally use to build app window. Basically you pass
-/// operation you want to happen on certain events here sort of like an observer/subscriber pattern. 
-/// I might turn this into that in the future.
-pub type CallOut = Box<dyn Fn() -> ()>;
-pub type TimedCallOut = Box<dyn Fn(std::time::Duration) -> ()>;
-pub type EventCallOut = Box<dyn Fn(&DeviceEvent) -> ()>;
-
-pub struct AppWindow{
-    pub state: graphics::State,
+///Combines the graphics device state and windowing into a single object.
+pub struct Graphics{
+    pub state: device::State,
     pub window: winit::window::Window,
 }
 
-impl AppWindow{
-    pub fn new(width: u32, height: u32, event_loop: &EventLoop<()>) -> Self{
+impl Graphics{
+    pub fn new(width: u32, height: u32, title: &str, event_loop: &EventLoop<()>) -> Self{
         let window = {
             let size = LogicalSize::new(width as f64, height as f64);
             WindowBuilder::new()
-                .with_title("Hello World")
+                .with_title(title)
                 .with_inner_size(size)
                 .with_min_inner_size(size)
                 .build(&event_loop)
                 .unwrap()
             };
-        
-        //Maybe have appwindow also return a &window and let the engine place the graphics state?
-        //or leave the window with a single graphics state? 
-        use futures::executor::block_on;
-        let state = block_on(graphics::State::new(&window));
 
-        return Self{
+        use futures::executor::block_on;
+        let state = block_on(device::State::new(&window));
+        
+        return Graphics{
             state,
             window,
         }
