+++ alchemy_framework/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod app_window;
 pub mod graphics;
 pub mod texture;
 pub mod camera;
